:truck: Move all api endpoints in single class
