+++ src/Unsplash.Tests/PhotosApiTests.cs
@@ -12,6 +12,7 @@ using WireMock.RequestBuilders;
 using WireMock.ResponseBuilders;
 using WireMock.Server;
 using Xunit;
+using static Unsplash.Api.ApiEndpoints;
 
 namespace Unsplash.Tests
 {
