+++ src/Unsplash.Tests/SearchApiTests.cs
@@ -8,6 +8,7 @@ using Unsplash.Models;
 using WireMock.ResponseBuilders;
 using WireMock.Server;
 using Xunit;
+using static Unsplash.Api.ApiEndpoints;
 
 namespace Unsplash.Tests
 {
