+++ src/Unsplash.Tests/StatsApiTests.cs
@@ -9,6 +9,7 @@ using Unsplash.Models;
 using WireMock.ResponseBuilders;
 using WireMock.Server;
 using Xunit;
+using static Unsplash.Api.ApiEndpoints;
 
 namespace Unsplash.Tests
 {
