+++ src/Unsplash.Tests/UsersApiTests.cs
@@ -9,6 +9,7 @@ using Unsplash.Client;
 using Unsplash.Models;
 using WireMock.ResponseBuilders;
 using Xunit;
+using static Unsplash.Api.ApiEndpoints;
 
 namespace Unsplash.Tests
 {
