+++ src/Unsplash/Api/Photos/PhotosApi.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Newtonsoft.Json;
 using Unsplash.Client;
 using Unsplash.Models;
+using static Unsplash.Api.ApiEndpoints;
 
 namespace Unsplash.Api.Photos
 {
