+++ src/Unsplash/Api/Search/SearchApi.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Unsplash.Client;
 using Unsplash.Extensions;
 using Unsplash.Models;
+using static Unsplash.Api.ApiEndpoints;
 
 namespace Unsplash.Api.Search
 {
