--- src/Unsplash/Api/Search/SearchApiUrls.cs
@@ -1,11 +1,0 @@
-﻿namespace Unsplash.Api.Search
-{
-    public static class SearchApiUrls
-    {
-        public static string Photos() => "/search/photos";
-
-        public static string Collections() => "/search/collections";
-
-        public static string Users() => "/search/users";
-    }
-}
