+++ src/Unsplash/Api/Users/UsersApi.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Unsplash.Client;
 using Unsplash.Extensions;
 using Unsplash.Models;
+using static Unsplash.Api.ApiEndpoints;
 
 namespace Unsplash.Api.Users
 {
